Refactor NX app bootstrap
Switch the v3 NX page to the new `NXProvider` wrapper and update the design system stylesheet import to the CSS entrypoint. Align integration tests with the legacy app entry and landing-page rendering, and refresh the NX icon asset.
