+++ apps/v3/app/NX/Theme/designSystem.ts
@@ -1,4 +1,4 @@
-import '@nx/design-system/styles';
+import '@nx/design-system/styles.css';
 
 export { Logo as NXLogo } from '@nx/design-system/components/brand/Logo.jsx';
